chore(editor): modify components
- add an editor component
- modify edit article component
- modify read article component
