+++ package.json
@@ -28,7 +28,7 @@
     "@ckeditor/ckeditor5-theme-lark": "^15.0.0",
     "@ckeditor/ckeditor5-upload": "^15.0.0",
     "ckeditor-cloudinary-uploader-adapter": "^1.1.3",
-    "ckeditor-code-snippet-plugin": "file:../packages/ckeditor-code-snippet-plugin/ckeditor-code-snippet-plugin-0.0.1.tgz",
+    "ckeditor5-code-snippet-plugin": "^1.0.2",
     "disqus-react": "^1.0.7",
     "dotenv-webpack": "^1.7.0",
     "prop-types": "^15.7.2",
