+++ src/config.js
@@ -9,7 +9,7 @@ const config = () => {
       };
     default:
       return {
-        backendUrl: 'https://iyikuyoro-staging.herokuapp.com/api/v1',
+        backendUrl: 'http://localhost:4000/api/v1',
         imageUploadPreset: 'example',
       };
   }
