+++ src/static/global.scss
@@ -7,3 +7,19 @@ body,
 html {
   margin: 0;
 }
+
+.code-snippet {
+  background-color: #1b1b1b;
+  color: #FFFFFF;
+  border-radius: 5px;
+  box-shadow: 1px 1px 11px grey;
+  padding: 10px;
+  max-width: 800px;
+  margin: 0 auto;
+  font-size: 0.7em;
+  overflow-x: scroll;
+}
+
+.ck .ck-editor__nested-editable.ck-editor__nested-editable_focused, .ck .ck-editor__nested-editable:focus {
+  color: black;
+}
